refactor: apply coding conformance fixes to lib/, test/, and headers
Engine Pimpl refactor — move callbacks, handlers, and data members from
Engine.h into Engine::Impl, hiding all implementation details.
IBackend interface slimmed — remove reset(), drain(),
setExclusiveMode(),
and isExclusiveMode() virtuals that were only used by PipeWire.
PipeWire callbacks restructured — convert anonymous namespace C callback
trampolines into Impl static member functions.
Coding conformance across lib/ and test/:
- Rule 4.2.1: add `final` to concrete Pimpl/helper structs
- Rule 4.3.1: add `const` to non-reassigned locals
- Rule 3.4.5: use `auto x = T{}` for non-primitive initialization
- Rule 3.1.7: use designated initializers for struct aggregation
- Rule 2.6.3: add `::` prefix to FLAC and LMDB C types
- Rule 2.1.2: add 29 missing blank lines around control blocks
- Rule 2.6.2: replace file-scope `static` with anonymous namespace
- Remove redundant `friend class` declarations in TrackView.h (C++11)
ClangTidy config — relaxed test checks inherit from strict with
selective
disables; add include/ to clang-tidy header filter.

+++ test/unit/audio/PcmConverterTest.cpp
@@ -23,7 +23,7 @@ TEST_CASE("PcmConverter: pad (Linear)", "[audio][pcm]")
     CHECK(destination[1] == -0x56780000);
     CHECK(destination[2] == 0x00000000);
     CHECK(destination[3] == 0x7FFF0000);
-    CHECK(destination[4] == -0x80000000);
+    CHECK(destination[4] == static_cast<std::int32_t>(-0x80000000));
   }
 
   SECTION("24-bit to 32-bit padding")
